Moved tests into its own folder
